Move the Queue components into the new Office scene.
