+++ src/scenes/MyMove/index.jsx
@@ -14,7 +14,6 @@ import { history } from 'shared/store';
 import Footer from 'shared/Footer';
 import Uploader from 'shared/Uploader';
 import PrivateRoute from 'shared/User/PrivateRoute';
-import Admin from 'scenes/Admin';
 
 import { getWorkflowRoutes } from './getWorkflowRoutes';
 import { createMove } from 'scenes/Moves/ducks';
@@ -50,7 +49,6 @@ export class AppWrapper extends Component {
               <Route path="/shipments/:shipmentsStatus" component={Shipments} />
               <Route path="/DD1299" component={DD1299} />
               <Route path="/feedback" component={Feedback} />
-              <Route path="/admin" component={Admin} />
               <PrivateRoute path="/upload" component={Uploader} />
               {getWorkflowRoutes(props)}
               <Route component={NoMatch} />
